[BE - feature] application 및 dto

+++ backend/board-service/src/main/java/com/a604/boardservice/entity/Message.java
@@ -18,7 +18,7 @@ public class Message {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Comment("쪽지 Seq")
-    private int seq;
+    private long seq;
 
     // 연결된 entity 추후 수정 필요
     @Column(nullable = false)
